update
Changelog:
 - Added aim bone option to soft aim and silent aim
 - Added color options to esp
 - Renamed "other" section in exploits to "gameplay"
 - Auto play is now inside the "gameplay" section (always in the exploits tab)
 - Added leave game
 - Updated pickaxe changer
 - Added force win
 - Added loadout level changer
 - Added add 10k fake gold
 - Added add 10k fake gems
 - Renamed "world" section to "game"  in exploits
 - Changed kill all method
 - Added watermark
 - Added 3D esp
 - Renamed soft aim to aimbot
 - Added locking modes to soft aim
 - Added bhop
 - Changed smoothing on soft aim
 - Added fun section to exploits
 - Fixed editing with silent aim

+++ Utils/Mouse.cs
@@ -8,7 +8,7 @@ namespace NetWare
         [DllImport("user32.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall)]
         private static extern void mouse_event(uint dwFlags, int dx, int dy, int dwData, int dwExtraInfo);
 
-        public static void MoveTo(Vector3 position, int lerp = 1)
+        public static void MoveTo(Vector3 position, float lerp = 1)
         {
             position.x = ((position.x - Render.screenCenter.x) / lerp);
             position.y = ((position.y - Render.screenCenter.y) / lerp);
